fix: Ajuste de segurança e exibição de informaçoes
